+++ sdk/formrecognizer/Azure.AI.FormRecognizer/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Release History
 
 ## 3.1.0-beta.1 (Unreleased)
+
+### Breaking changes
 - It defaults to the latest supported API version, which currently is `2.1-preview.1`.
-Note that new functionality hasn't been implemented in the client library.
+
+### New Features
+- Added ability to create a composed model from the `FormTrainingClient`.
+- Added the properties `DisplayName` and `Properties` to types `CustomFormModel` and `CustomFormModelInfo`.
+- Added property `DisplayName` to `TrainingOptions` and new type `CreateComposedModelOptions`.
+- Added type `CustomFormModelProperties` that includes information like if a model is a composed model.
+- Added property `ModelId` to `CustomFormSubmodel` and `TrainingDocumentInfo`.
+- Added properties `ModelId` and `FormTypeConfidence` to `RecognizedForm`.
 
 ## 3.0.0 (2020-08-20)